+++ src/strauss/__init__.py
@@ -60,3 +60,7 @@ def get_table(name=None, source=None, include_input=False):
 def list_tables():
     fig = _get_current_figure()
     fig.list_tables()
+
+def plot_mapping(name=None, **kwargs):
+    fig = _get_current_figure()
+    return fig.plot_mapping(name, **kwargs)